services/svc-fusion-extension/app/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
asyncpg==0.29.0
httpx==0.27.2
tenacity==9.0.0
pydantic==2.10.4
pydantic-settings==2.7.0
PyJWT==2.9.0
azure-storage-blob==12.23.0
azure-core==1.32.0